@@ -1,12 +1,13 @@
-/*
- * Bedrock webpack Module Configuration
- *
- * Copyright (c) 2017-2018 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2017-2022 Digital Bazaar, Inc. All rights reserved.
  */
-const bedrock = require('bedrock');
+import * as bedrock from 'bedrock';
+import {fileURLToPath} from 'url';
+import path from 'path';
+
 const cc = bedrock.util.config.main.computer();
-const config = bedrock.config;
-const path = require('path');
+const {config} = bedrock;
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 config['bedrock-webpack'] = {};
 
